requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
matplotlib
reportlab
pypdf
python-docx

